Consolidate to AdminController if uploadType is set
