+++ src/AdminController.js
@@ -2,6 +2,9 @@ import {EditController} from 'nxus-web'
 import {nav} from 'nxus-web'
 import {actions} from 'nxus-web'
 import {permissions} from 'nxus-users'
+import {templater} from 'nxus-templater'
+import {dataManager} from 'nxus-data-manager'
+import {admin} from './'
 import morph from 'morph'
 
 import {application as app} from 'nxus-core'
@@ -16,10 +19,14 @@ class AdminController extends EditController {
 
     super(opts)
 
+
     this.displayName = opts.displayName || morph.toTitle(this.modelIdentity)
     this.icon = opts.icon || 'fa fa-files-o'
     this.order = opts.order || 0
 
+    this.uploadOptions = opts.uploadOptions || {}
+    this.uploadType = opts.uploadType || null
+    
     if(this.routePrefix[0] != '/') this.routePrefix = '/'+this.routePrefix
     
     nav.add('admin-sidebar', this.displayName, this.routePrefix, {subMenu: this.prefix+'-submenu', icon: this.icon, order: this.order})
@@ -46,7 +53,47 @@ class AdminController extends EditController {
       null,
       "Admin"
     )
-  } 
+
+    if (this.uploadType) {
+      this._setupImport()
+    }
+  }
+
+  _setupImport() {
+    let importRoute = this.routePrefix+'/import'
+    dataManager.uploadPath(importRoute, 'file')
+    admin.page({
+      route: importRoute,
+      nav: false,
+      upload: this.uploadType,
+      base: this.routePrefix,
+      title: "Import "+this.displayName,
+      displayName: this.displayName
+    }, this.templatePrefix+'-import')
+    admin.page({
+      route: importRoute,
+      nav: false,
+      directHandler: true,
+      method: 'POST'
+    }, ::this._saveImport)
+    
+    templater.default().template(__dirname+"/templates/admin-import.ejs", this.pageTemplate, this.templatePrefix+"-import")
+
+    nav.add(this.prefix+'-submenu', 'Import', importRoute, {icon: 'fa fa-upload'})
+    actions.add(this.templatePrefix+"-list", "Import", "/import", {
+      icon: "fa fa-upload"
+    })
+    
+  }
+  
+  _saveImport (req, res) {
+    let opts = {type: this.uploadType, ...this.uploadOptions}
+    dataManager.importFileToModel(this.modelIdentity, req.file.path, opts)
+    .then((insts) => {
+      req.flash('info', 'Imported '+insts.length+' '+this.displayName)
+      res.redirect(this.routePrefix)
+    })
+  }
 }
 
 export default AdminController
